refactor to separate env code from lqr code
