+++ .gitignore
@@ -1,1 +1,3 @@
 runs
+__pycache__
+*.egg-info
